[TECH](RC) renamed some files to match nomenclature with -
