+++ package.json
@@ -39,8 +39,8 @@
     "create-swagger": "node --async-stack-traces ./create-swagger.js",
     "create-swagger:dev": "tsx ./src/create-swagger.ts",
     "------------------ TASKS -----------------": "",
-    "extract-data-from-fds": "LOG_LEVEL=error tsx ./src/tasks/poc/extract_data_from_fds.ts",
-    "extract-data-from-all-fds": "LOG_LEVEL=error tsx ./src/tasks/poc/extract_data_from_all_fds.ts",
+    "extract-data-from-fds": "LOG_LEVEL=error tsx ./src/tasks/extract/from-fds.ts",
+    "extract-data-from-all-fds": "LOG_LEVEL=error tsx ./src/tasks/extract/from-all-fds.ts",
     "------------------ TEST ------------------": "",
     "test": "NODE_ENV=test npm run init-db && npx vitest run --no-threads --no-isolate --coverage",
     "jest:circleci": "LOG_LEVEL=error npx vitest run --no-threads --no-isolate --coverage --reporter junit --outputFile reports/test-results.xml",
