+++ src/routes.ts
@@ -1,7 +1,7 @@
 import { Router } from 'express';
 
 import { healthCheckController } from '@topics/health-check/health-check.controller.js';
-import { fdsEngineController } from '@topics/engine/fds_engine.controller.js';
+import { fdsEngineController } from '@topics/engine/fds-engine.controller.js';
 
 const controllers = [healthCheckController, fdsEngineController];
